Title: Fixed a bug causing wrong color changes.
Description:
Fixed a bug which was causing unwanted color
changes to other buffers when any buffer had any
change in the color of colorcolumn. So basically
the bug was changing the color of colorcolumn
globally instead of the active buffer. Also,
simplified the configurations, updated the
default configurations and renamed the file as
well.

CID: #3